+++ src/som/primitives/HashPrim.java
@@ -7,12 +7,12 @@ import com.oracle.truffle.api.dsl.Specialization;
 
 public abstract class HashPrim extends UnarySideEffectFreeExpressionNode {
   @Specialization
-  public final int doSString(final String receiver) {
+  public final long doSString(final String receiver) {
     return receiver.hashCode();
   }
 
   @Specialization
-  public final int doSAbstractObject(final SAbstractObject receiver) {
+  public final long doSAbstractObject(final SAbstractObject receiver) {
     return receiver.hashCode();
   }
 }
