+++ src/som/primitives/LengthPrim.java
@@ -6,12 +6,12 @@ import com.oracle.truffle.api.dsl.Specialization;
 
 public abstract class LengthPrim extends UnarySideEffectFreeExpressionNode {
   @Specialization
-  public final int doSArray(final Object[] receiver) {
+  public final long doSArray(final Object[] receiver) {
     return receiver.length;
   }
 
   @Specialization
-  public final int doSString(final String receiver) {
+  public final long doSString(final String receiver) {
     return receiver.length();
   }
 }
